ai_chat/requirements.txt
fastapi
uvicorn
python-dotenv
groq
elevenlabs
langchain
langchain-google-genai
langchain-community
langgraph
redis
pydantic
python-multipart
langchain_tavily